+++ server/src/routes/userRoutes.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import {userController} from '../controller/controller';
+import {userController} from '../controller/userController';
 
 const userRouter = express.Router();
 
@@ -8,7 +8,7 @@ userRouter.get('/', userController.getHomePage);
 
 userRouter.get('/user/:id',userController.getUserById)
 
-userRouter.post('/reg',userController.register);
+userRouter.post('/reg',userController.reg);
